@@ -13,7 +13,7 @@ namespace UndertaleModLib.Decompiler
 {
     public static class Assembler
     {
-        public static Dictionary<short, string> BreakIDToName = new Dictionary<short, string>()
+        internal static readonly Dictionary<short, string> BreakIDToName = new()
         {
             { -1,  "chkindex" },
             { -2,  "pushaf" },
@@ -27,7 +27,7 @@ namespace UndertaleModLib.Decompiler
             { -10, "chknullish" },
             { -11, "pushref" }
         };
-        public static Dictionary<string, short> NameToBreakID = new Dictionary<string, short>()
+        internal static readonly Dictionary<string, short> NameToBreakID = new()
         {
             { "chkindex", -1 },
             { "pushaf", -2 },
@@ -130,8 +130,7 @@ namespace UndertaleModLib.Decompiler
                         if (line == "<drop>")
                         {
                             instr.JumpOffsetPopenvExitMagic = true;
-                            if (data?.GeneralInfo?.BytecodeVersion <= 14)
-                                instr.JumpOffset = -1048576; // I really don't know at this point. Magic for little endian 00 00 F0
+                            instr.JumpOffset = 0xF00000;
                         }
                         else if (line[0] == '[' && line[^1] == ']')
                             label = line.Substring(1, line.Length - 2);
@@ -151,7 +150,7 @@ namespace UndertaleModLib.Decompiler
                     else
                     {
                         UndertaleInstruction.InstanceType inst = instr.TypeInst;
-                        instr.Destination = ParseVariableReference(line, vars, localvars, ref inst, instr, data);
+                        (instr.ValueVariable, instr.ReferenceType) = ParseVariableReference(line, vars, localvars, ref inst, instr, data);
                         instr.TypeInst = inst;
                     }
                     line = "";
@@ -161,63 +160,70 @@ namespace UndertaleModLib.Decompiler
                     switch (instr.Type1)
                     {
                         case UndertaleInstruction.DataType.Double:
-                            instr.Value = Double.Parse(line, CultureInfo.InvariantCulture);
-                            break;
-                        case UndertaleInstruction.DataType.Float:
-                            instr.Value = Single.Parse(line, CultureInfo.InvariantCulture);
+                            instr.ValueDouble = double.Parse(line, CultureInfo.InvariantCulture);
                             break;
                         case UndertaleInstruction.DataType.Int32:
-                            int ival;
-                            if (Int32.TryParse(line, out ival))
-                                instr.Value = ival;
+                            if (int.TryParse(line, out int ival))
+                            {
+                                instr.ValueInt = ival;
+                            }
                             else
                             {
-                                if (line.StartsWith("[variable]"))
+                                if (line.StartsWith("[variable]", StringComparison.Ordinal))
                                 {
-                                    line = line.Substring("[variable]".Length);
-                                    instr.Value = new UndertaleInstruction.Reference<UndertaleVariable>(data.Variables.EnsureDefined(line,
-                                        UndertaleInstruction.InstanceType.Self, false, data.Strings, data));
+                                    line = line["[variable]".Length..];
+                                    instr.ValueVariable = data.Variables.EnsureDefined(
+                                        data.Strings.MakeString(line, out int nameStringId), nameStringId,
+                                        UndertaleInstruction.InstanceType.Self, false, data);
                                 }
-                                else if (line.StartsWith("[function]"))
+                                else if (line.StartsWith("[function]", StringComparison.Ordinal))
                                 {
-                                    line = line.Substring("[function]".Length);
-                                    instr.Value = new UndertaleInstruction.Reference<UndertaleFunction>(data.Functions.ByName(line));
+                                    line = line["[function]".Length..];
+                                    instr.ValueFunction = data.Functions.ByName(line);
                                 }
                                 else
                                 {
-                                    var f = data.Functions.ByName(line);
-                                    if (f == null)
-                                        instr.Value = (int)ParseResourceName(line, data);
+                                    if (data.Functions.ByName(line) is UndertaleFunction f)
+                                    {
+                                        instr.ValueFunction = f;
+                                    }
                                     else
-                                        instr.Value = new UndertaleInstruction.Reference<UndertaleFunction>(f);
+                                    {
+                                        instr.ValueInt = ParseResourceName(line, data);
+                                    }
                                 }
                             }
                             break;
                         case UndertaleInstruction.DataType.Int64:
-                            long lval;
-                            if (Int64.TryParse(line, out lval))
-                                instr.Value = lval;
+                            if (long.TryParse(line, out long lval))
+                            {
+                                instr.ValueLong = lval;
+                            }
                             else
-                                instr.Value = (long)ParseResourceName(line, data);
-                            break;
-                        case UndertaleInstruction.DataType.Boolean:
-                            instr.Value = bool.Parse(line);
+                            {
+                                instr.ValueLong = ParseResourceName(line, data);
+                            }
                             break;
                         case UndertaleInstruction.DataType.Variable:
                             UndertaleInstruction.InstanceType inst2 = instr.TypeInst;
-                            instr.Value = ParseVariableReference(line, vars, localvars, ref inst2, instr, data);
+                            (instr.ValueVariable, instr.ReferenceType) = ParseVariableReference(line, vars, localvars, ref inst2, instr, data);
                             instr.TypeInst = inst2;
                             break;
                         case UndertaleInstruction.DataType.String:
-                            instr.Value = ParseStringReference(line, strg);
+                            instr.ValueString = ParseStringReference(line, strg);
                             break;
                         case UndertaleInstruction.DataType.Int16:
-                            short sval;
-                            if (Int16.TryParse(line, out sval))
-                                instr.Value = sval;
+                            if (short.TryParse(line, out short sval))
+                            {
+                                instr.ValueShort = sval;
+                            }
                             else
-                                instr.Value = (short)ParseResourceName(line, data);
+                            {
+                                instr.ValueShort = (short)ParseResourceName(line, data);
+                            }
                             break;
+                        default:
+                            throw new Exception("Invalid push data type");
                     }
                     line = "";
                     break;
@@ -230,7 +236,7 @@ namespace UndertaleModLib.Decompiler
                     UndertaleFunction func = funcs.ByName(match.Groups[1].Value);
                     if (func == null)
                         throw new Exception("Function not found: " + match.Groups[1].Value);
-                    instr.Function = new UndertaleInstruction.Reference<UndertaleFunction>() { Target = func };
+                    instr.ValueFunction = func;
                     instr.ArgumentsCount = UInt16.Parse(match.Groups[2].Value);
                     line = "";
                     break;
@@ -238,11 +244,11 @@ namespace UndertaleModLib.Decompiler
                 case UndertaleInstruction.InstructionType.BreakInstruction:
                     if (breakId != 0)
                     {
-                        instr.Value = breakId;
+                        instr.ExtendedKind = breakId;
                         if (breakId == -11) // pushref
                         {
                             // Parse additional int argument
-                            if (Int32.TryParse(line, out int intArgument))
+                            if (int.TryParse(line, out int intArgument))
                             {
                                 instr.IntArgument = intArgument;
                             }
@@ -252,12 +258,12 @@ namespace UndertaleModLib.Decompiler
                                 var f = data.Functions.ByName(line);
                                 if (f == null)
                                     throw new Exception("Function in pushref not found: " + line);
-                                instr.Function = new UndertaleInstruction.Reference<UndertaleFunction>(f);
+                                instr.ValueFunction = f;
                             }
                         }
                     }
                     else
-                        instr.Value = Int16.Parse(line);
+                        instr.ExtendedKind = Int16.Parse(line);
                     line = "";
                     break;
             }
@@ -284,11 +290,13 @@ namespace UndertaleModLib.Decompiler
         public static List<UndertaleInstruction> Assemble(string source, IList<UndertaleFunction> funcs, IList<UndertaleVariable> vars, IList<UndertaleString> strg, UndertaleData data = null)
         {
             StringReader strReader = new(source);
-            uint addr = 0;
-            Dictionary<string, uint> labels = new Dictionary<string, uint>();
-            Dictionary<UndertaleInstruction, string> labelTargets = new Dictionary<UndertaleInstruction, string>();
-            List<UndertaleInstruction> instructions = new List<UndertaleInstruction>();
-            Dictionary<string, UndertaleVariable> localvars = new Dictionary<string, UndertaleVariable>();
+
+            Dictionary<string, uint> labels = new();
+            List<(UndertaleInstruction Instruction, uint InstructionAddress, string Label)> labelTargets = new();
+            List<UndertaleInstruction> instructions = new(16);
+            Dictionary<string, UndertaleVariable> localvars = new();
+
+            uint address = 0;
             string fullLine;
             while ((fullLine = strReader.ReadLine()) is not null)
             {
@@ -304,9 +312,7 @@ namespace UndertaleModLib.Decompiler
                     line = line.Substring(2, line.Length - 2).Trim();
                     int space = line.IndexOf(' ', StringComparison.InvariantCulture);
                     string codeName = line.Substring(0, space);
-                    var code = data.Code.ByName(codeName);
-                    if (code == null)
-                        throw new Exception($"Failed to find code entry with name \"{codeName}\".");
+                    UndertaleCode code = data.Code.ByName(codeName) ?? throw new Exception($"Failed to find code entry with name \"{codeName}\".");
                     string info = line.Substring(space + 1);
 
                     Match match = codeEntryRegex.Match(info);
@@ -314,7 +320,7 @@ namespace UndertaleModLib.Decompiler
                         throw new Exception("Sub-code entry format error");
                     code.LocalsCount = ushort.Parse(match.Groups[1].Value);
                     code.ArgumentsCount = ushort.Parse(match.Groups[2].Value);
-                    code.Offset = addr * 4;
+                    code.Offset = address * 4;
                     continue;
                 }
                 if (line[0] == ':' && line.Length >= 3)
@@ -327,7 +333,7 @@ namespace UndertaleModLib.Decompiler
                         {
                             if (labels.ContainsKey(label))
                                 throw new Exception("Duplicate label: " + label);
-                            labels.Add(label, addr);
+                            labels.Add(label, address);
                         }
 
                         continue;
@@ -360,17 +366,16 @@ namespace UndertaleModLib.Decompiler
 
                 string labelTgt;
                 UndertaleInstruction instr = AssembleOne(line, funcs, vars, strg, localvars, out labelTgt, data);
-                instr.Address = addr;
                 if (labelTgt != null)
-                    labelTargets.Add(instr, labelTgt);
+                    labelTargets.Add((instr, address, labelTgt));
 
                 instructions.Add(instr);
 
-                addr += instr.CalculateInstructionSize();
+                address += instr.CalculateInstructionSize();
             }
-            foreach (var pair in labelTargets)
+            foreach ((UndertaleInstruction instr, uint instrAddress, string label) in labelTargets)
             {
-                pair.Key.JumpOffset = (int)labels[pair.Value] - (int)pair.Key.Address;
+                instr.JumpOffset = (int)labels[label] - (int)instrAddress;
             }
             return instructions;
         }
@@ -416,7 +421,8 @@ namespace UndertaleModLib.Decompiler
             return new UndertaleResourceById<UndertaleString, UndertaleChunkSTRG>() { Resource = strobj, CachedId = (int)id.Value };
         }
 
-        private static UndertaleInstruction.Reference<UndertaleVariable> ParseVariableReference(string line, IList<UndertaleVariable> vars, Dictionary<string, UndertaleVariable> localvars, ref UndertaleInstruction.InstanceType instance, UndertaleInstruction instr, UndertaleData data = null)
+        private static (UndertaleVariable Variable, UndertaleInstruction.VariableType ReferenceType) ParseVariableReference(
+            string line, IList<UndertaleVariable> vars, Dictionary<string, UndertaleVariable> localvars, ref UndertaleInstruction.InstanceType instance, UndertaleInstruction instr, UndertaleData data = null)
         {
             ReadOnlySpan<char> str = line.AsSpan();
             int strPosition = 0;
@@ -526,9 +532,9 @@ namespace UndertaleModLib.Decompiler
             // Locate variable from either local variables, or VARI chunk
             UndertaleVariable locatedVariable;
             string variableName = str[strPosition..].ToString();
-            if (variInstanceType == UndertaleInstruction.InstanceType.Local)
+            if (variInstanceType == UndertaleInstruction.InstanceType.Local && data?.CodeLocals is not null)
             {
-                locatedVariable = localvars.ContainsKey(variableName) ? localvars[variableName] : null;
+                locatedVariable = localvars.TryGetValue(variableName, out UndertaleVariable value) ? value : null;
             }
             else
             {
@@ -542,7 +548,7 @@ namespace UndertaleModLib.Decompiler
             }
 
             // Return reference to be used in instruction
-            return new UndertaleInstruction.Reference<UndertaleVariable>(locatedVariable, type);
+            return (locatedVariable, type);
         }
     }
 }
