@@ -11,14 +11,9 @@ namespace UndertaleModLib.Decompiler
     {
         private static void AppendLocalVarDefinitionsToStringBuilder(StringBuilder sb, UndertaleCode code, IList<UndertaleVariable> vars, UndertaleCodeLocals locals)
         {
-            if (code.WeirdLocalFlag && locals is null)
-            {
-                return;
-            }
-
             if (locals is null)
             {
-                sb.AppendLine("; Missing code locals, possibly due to unsupported bytecode version or brand new code entry.");
+                sb.Append("; Missing code locals, possibly due to unsupported bytecode version or brand new code entry.\n");
                 return;
             }
             
@@ -37,64 +32,108 @@ namespace UndertaleModLib.Decompiler
                     sb.Append(' ');
                     sb.Append(vars.IndexOf(refVar));
                 }
-                sb.AppendLine();
+                sb.Append('\n');
             }
         }
         
-        public static string Disassemble(this UndertaleCode code, IList<UndertaleVariable> vars, UndertaleCodeLocals locals)
+        public static string Disassemble(this UndertaleCode code, IList<UndertaleVariable> vars, UndertaleCodeLocals locals, bool ignoreMissingCodeLocals = false)
         {
             // This StringBuilder is shared with the ToString method of the code instructions.
-            // Experimentation has shown that 200 is a good enough starting value for it. 
+            // Experimentation has shown that 200 is a good enough starting value for it (now changed to 256).
             // 300 seemed too high and 100 too low. This may change in the future.
-            StringBuilder sb = new StringBuilder(200);
-            if (locals is null && !code.WeirdLocalFlag)
-                sb.AppendLine("; WARNING: Missing code locals, possibly due to unsupported bytecode version or a brand new code entry.");
-            else
-                AppendLocalVarDefinitionsToStringBuilder(sb, code, vars, locals);
+            StringBuilder sb = new(256);
 
+            // Print locals
+            bool printedLocalVars = true;
+            if (locals is null && !code.WeirdLocalFlag)
+            {
+                if (ignoreMissingCodeLocals)
+                {
+                    printedLocalVars = false;
+                }
+                else
+                {
+                    sb.Append("; WARNING: Missing code locals, possibly due to unsupported bytecode version or a brand new code entry.\n");
+                }
+            }
+            else if (locals is null && code.WeirdLocalFlag)
+            {
+                printedLocalVars = false;
+            }
+            else
+            {
+                AppendLocalVarDefinitionsToStringBuilder(sb, code, vars, locals);
+            }
+
+            // Collect fragments (sub-functions/child code entries)
             Dictionary<uint, string> fragments = new(code.ChildEntries.Count);
             foreach (var dup in code.ChildEntries)
             {
                 fragments.Add(dup.Offset / 4, $"{(dup.Name?.Content ?? "<null>")} (locals={dup.LocalsCount}, argc={dup.ArgumentsCount})");
             }
 
-            List<uint> blocks = FindBlockAddresses(code);
+            // Find addresses of all blocks
+            Dictionary<uint, uint> blocks = FindBlockAddresses(code);
 
-            foreach (var inst in code.Instructions)
+            // Print actual instructions
+            uint address = 0;
+            foreach (UndertaleInstruction inst in code.Instructions)
             {
-                bool doNewline = true;
-                if (fragments.TryGetValue(inst.Address, out string entry))
+                // Print an extra newline for fragments/blocks if local vars were printed at the beginning,
+                // or if any instructions have already been printed.
+                bool doNewline = printedLocalVars || address > 0;
+
+                // Print fragment at current address
+                if (fragments.TryGetValue(address, out string entry))
                 {
-                    sb.AppendLine();
-                    sb.AppendLine($"> {entry}");
+                    sb.Append('\n');
+                    sb.Append($"> {entry}\n");
+
+                    // No need for a second newline before blocks
                     doNewline = false;
                 }
 
-                int ind = blocks.IndexOf(inst.Address);
-                if (ind != -1)
+                // Print block at current address
+                if (blocks.TryGetValue(address, out uint ind))
                 {
                     if (doNewline)
-                        sb.AppendLine();
-                    sb.AppendLine($":[{ind}]");
+                    {
+                        sb.Append('\n');
+                    }
+                    sb.Append($":[{ind}]\n");
                 }
 
-                inst.ToString(sb, code, blocks);
-                sb.AppendLine();
+                // Print actual instruction at current address
+                inst.ToString(sb, code, address, blocks);
+                sb.Append('\n');
+
+                // Advance address to next instruction
+                address += inst.CalculateInstructionSize();
             }
 
-            sb.AppendLine();
+            // Print ending block
+            if (printedLocalVars || address > 0)
+            {
+                sb.Append('\n');
+            }
             sb.Append(":[end]");
 
             return sb.ToString();
         }
 
-        public static List<uint> FindBlockAddresses(UndertaleCode code)
+        public static Dictionary<uint, uint> FindBlockAddresses(UndertaleCode code)
         {
-            HashSet<uint> addresses = new HashSet<uint>();
+            // Use a sorted set, so that block indices can be calculated
+            SortedSet<uint> addresses = new();
 
-            if (code.Instructions.Count != 0)
+            // Add initial block, if any instructions exist
+            if (code.Instructions.Count > 0)
+            {
                 addresses.Add(0);
+            }
 
+            // Add all other blocks based 
+            uint currentAddress = 0;
             foreach (var inst in code.Instructions)
             {
                 switch (inst.Kind)
@@ -103,23 +142,32 @@ namespace UndertaleModLib.Decompiler
                     case UndertaleInstruction.Opcode.Bf:
                     case UndertaleInstruction.Opcode.Bt:
                     case UndertaleInstruction.Opcode.PushEnv:
-                        addresses.Add(inst.Address + 1);
-                        addresses.Add((uint)(inst.Address + inst.JumpOffset));
+                        addresses.Add(currentAddress + 1);
+                        addresses.Add((uint)(currentAddress + inst.JumpOffset));
                         break;
                     case UndertaleInstruction.Opcode.PopEnv:
                         if (!inst.JumpOffsetPopenvExitMagic)
-                            addresses.Add((uint)(inst.Address + inst.JumpOffset));
+                        {
+                            addresses.Add((uint)(currentAddress + inst.JumpOffset));
+                        }
                         break;
                     case UndertaleInstruction.Opcode.Exit:
                     case UndertaleInstruction.Opcode.Ret:
-                        addresses.Add(inst.Address + 1);
+                        addresses.Add(currentAddress + 1);
                         break;
                 }
+                currentAddress += inst.CalculateInstructionSize();
             }
 
-            List<uint> res = addresses.ToList();
-            res.Sort();
-            return res;
+            // Convert to an index lookup
+            Dictionary<uint, uint> blockIndicesByAddress = new(addresses.Count);
+            uint blockIndex = 0;
+            foreach (uint address in addresses)
+            {
+                blockIndicesByAddress.Add(address, blockIndex);
+                blockIndex++;
+            }
+            return blockIndicesByAddress;
         }
     }
 }
