@@ -30,6 +30,7 @@ using System.Windows.Media.Animation;
 using System.Windows.Media.Imaging;
 using System.Windows.Media.TextFormatting;
 using System.Windows.Navigation;
+using System.Windows.Shell;
 using System.Xml;
 using UndertaleModLib;
 using UndertaleModLib.Compiler;
@@ -288,8 +289,6 @@ namespace UndertaleModTool
         private async void TabControl_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             UndertaleCode code = this.DataContext as UndertaleCode;
-            Directory.CreateDirectory(MainPath);
-            Directory.CreateDirectory(TempPath);
             if (code == null)
                 return;
 
@@ -503,7 +502,7 @@ namespace UndertaleModTool
                 foreach (var obj in list)
                 {
                     if (obj is not UndertaleNamedResource namedObj)
-                        return;
+                        continue;
 
                     NamedObjDict[namedObj.Name.Content] = namedObj;
                 }
@@ -533,8 +532,6 @@ namespace UndertaleModTool
 
         private void DisassembleCode(UndertaleCode code, bool first)
         {
-            code.UpdateAddresses();
-
             string text;
 
             int currLine = 1;
@@ -560,7 +557,8 @@ namespace UndertaleModTool
             if (code.ParentEntry != null)
             {
                 DisassemblyEditor.IsReadOnly = true;
-                text = "; This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there";
+                text = "; This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there.";
+                DisassemblyChanged = false;
             }
             else
             {
@@ -569,7 +567,7 @@ namespace UndertaleModTool
                 try
                 {
                     var data = mainWindow.Data;
-                    text = code.Disassemble(data.Variables, data.CodeLocals.For(code));
+                    text = code.Disassemble(data.Variables, data.CodeLocals?.For(code), data.CodeLocals is null);
 
                     CurrentLocals.Clear();
                 }
@@ -604,14 +602,14 @@ namespace UndertaleModTool
             gettext = new Dictionary<string, string>();
             string[] decompilationOutput;
             GlobalDecompileContext context = new(data);
-            if (!SettingsWindow.ProfileModeEnabled)
+            if (!SettingsWindow.ProfileModeEnabled || mainWindow.ProfileHash is not string currentMD5)
             {
                 decompilationOutput = new Underanalyzer.Decompiler.DecompileContext(context, gettextCode, data.ToolInfo.DecompilerSettings)
                     .DecompileToString().Split('\n');
             }
             else
             {
-                string path = Path.Combine(TempPath, gettextCode.Name.Content + ".gml");
+                string path = Path.Combine(MainWindow.ProfilesFolder, currentMD5, "Temp", gettextCode.Name.Content + ".gml");
                 if (File.Exists(path))
                 {
                     try
@@ -630,7 +628,8 @@ namespace UndertaleModTool
                         .DecompileToString().Split('\n');
                 }
             }
-            Regex textdataRegex = new Regex("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", \\\"(.*)\\\"\\)", RegexOptions.Compiled);
+            Regex textdataRegex = new("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", \\\"(.*)\\\"\\)", RegexOptions.Compiled);
+            Regex textdataRegex2 = new("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", '(.*)'\\)", RegexOptions.Compiled);
             foreach (var line in decompilationOutput)
             {
                 Match m = textdataRegex.Match(line);
@@ -638,7 +637,14 @@ namespace UndertaleModTool
                 {
                     try
                     {
-                        gettext.Add(m.Groups[1].Value, m.Groups[2].Value);
+                        if (!data.IsGameMaker2() && m.Groups[2].Value.Contains("'\"'"))
+                        {
+                            gettext.Add(m.Groups[1].Value, $"\"{m.Groups[2].Value}\"");
+                        }
+                        else
+                        {
+                            gettext.Add(m.Groups[1].Value, m.Groups[2].Value);
+                        }
                     }
                     catch (ArgumentException)
                     {
@@ -647,6 +653,32 @@ namespace UndertaleModTool
                     catch
                     {
                         mainWindow.ShowError("Unknown error in textdata_en. This may cause errors in the comment display of text.");
+                    }
+                }
+                else
+                {
+                    m = textdataRegex2.Match(line);
+                    if (m.Success)
+                    {
+                        try
+                        {
+                            if (!data.IsGameMaker2() && m.Groups[2].Value.Contains("\"'\""))
+                            {
+                                gettext.Add(m.Groups[1].Value, $"\'{m.Groups[2].Value}\'");
+                            }
+                            else
+                            {
+                                gettext.Add(m.Groups[1].Value, m.Groups[2].Value);
+                            }
+                        }
+                        catch (ArgumentException)
+                        {
+                            mainWindow.ShowError("There is a duplicate key in textdata_en, being " + m.Groups[1].Value + ". This may cause errors in the comment display of text.");
+                        }
+                        catch
+                        {
+                            mainWindow.ShowError("Unknown error in textdata_en. This may cause errors in the comment display of text.");
+                        }
                     }
                 }
             }
@@ -696,7 +728,8 @@ namespace UndertaleModTool
 
             if (code.ParentEntry != null)
             {
-                DecompiledEditor.Text = "// This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there";
+                DecompiledEditor.Text = "// This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there.";
+                DecompiledChanged = false;
                 CurrentDecompiled = code;
                 existingDialog?.TryClose();
             }
@@ -745,15 +778,23 @@ namespace UndertaleModTool
                     Exception e = null;
                     try
                     {
-                        string path = Path.Combine(TempPath, code.Name.Content + ".gml");
-                        if (!SettingsWindow.ProfileModeEnabled || !File.Exists(path))
+                        if (!SettingsWindow.ProfileModeEnabled || mainWindow.ProfileHash is not string currentMD5)
                         {
                             decompiled = new Underanalyzer.Decompiler.DecompileContext(context, code, dataa.ToolInfo.DecompilerSettings)
                                 .DecompileToString();
                         }
                         else
                         {
-                            decompiled = File.ReadAllText(path);
+                            string path = Path.Combine(MainWindow.ProfilesFolder, currentMD5, "Temp", code.Name.Content + ".gml");
+                            if (!File.Exists(path))
+                            {
+                                decompiled = new Underanalyzer.Decompiler.DecompileContext(context, code, dataa.ToolInfo.DecompilerSettings)
+                                    .DecompileToString();
+                            }
+                            else
+                            {
+                                decompiled = File.ReadAllText(path);
+                            }
                         }
                     }
                     catch (Exception ex)
@@ -849,9 +890,6 @@ namespace UndertaleModTool
 
                             if (existingDialog is not null)                      //if code was edited (and compiles after it)
                             {
-                                dataa.GMLCacheChanged.Add(code.Name.Content);
-                                dataa.GMLCacheFailed?.Remove(code.Name.Content); //remove that code name, since that code compiles now
-
                                 openSaveDialog = mainWindow.IsSaving;
                             }
                         }
@@ -882,7 +920,7 @@ namespace UndertaleModTool
             CurrentLocals.Clear();
 
             // Look up locals for given code entry's name, for syntax highlighting
-            var locals = data.CodeLocals.ByName(code.Name.Content);
+            var locals = data.CodeLocals?.ByName(code.Name.Content);
             if (locals != null)
             {
                 foreach (var local in locals.Locals)
@@ -954,68 +992,52 @@ namespace UndertaleModTool
                 // This is still a problem in rare cases for some unknown reason
             }
 
-            CompileContext compileContext = null;
+            CompileResult compileResult = new();
+            string rootException = null;
             string text = DecompiledEditor.Text;
             var dispatcher = Dispatcher;
             Task t = Task.Run(() =>
             {
                 try
                 {
-                    compileContext = Compiler.CompileGMLText(text, data, code, (f) => { dispatcher.Invoke(() => f()); });
+                    CompileGroup group = new(data);
+                    group.MainThreadAction = (f) => { dispatcher.Invoke(() => f()); };
+                    group.QueueCodeReplace(code, text);
+                    compileResult = group.Compile();
                 }
                 catch (Exception ex)
                 {
-                    compileContext = new(data, code)
-                    {
-                        HasError = true,
-                        ResultError = ex.ToString()
-                    };
+                    rootException = ex.ToString();
                 }
             });
             await t;
 
-            if (compileContext == null)
+            if (rootException is not null)
             {
                 dialog.TryClose();
-                mainWindow.ShowError("Compile context was null for some reason...", "This shouldn't happen");
+                mainWindow.ShowError(Truncate(rootException, 512), "Compiler error");
                 return;
             }
 
-            if (compileContext.HasError)
+            if (!compileResult.Successful)
             {
                 dialog.TryClose();
-                mainWindow.ShowError(Truncate(compileContext.ResultError, 512), "Compiler error");
+                mainWindow.ShowError(Truncate(compileResult.PrintAllErrors(false), 512), "Compiler error");
                 return;
             }
 
-            if (!compileContext.SuccessfulCompile)
+            if (SettingsWindow.ProfileModeEnabled && mainWindow.ProfileHash is string currentMD5)
             {
-                dialog.TryClose();
-                mainWindow.ShowError("(unknown error message)", "Compile failed");
-                return;
-            }
-
-            code.Replace(compileContext.ResultAssembly);
-            try
-            {
-                string path = Path.Combine(TempPath, code.Name.Content + ".gml");
-                if (SettingsWindow.ProfileModeEnabled)
+                try
                 {
                     // Write text, only if in the profile mode.
+                    string path = Path.Combine(MainWindow.ProfilesFolder, currentMD5, "Temp", code.Name.Content + ".gml");
                     File.WriteAllText(path, DecompiledEditor.Text);
                 }
-                else
+                catch (Exception exc)
                 {
-                    // Destroy file with comments if it's been edited outside the profile mode.
-                    // We're dealing with the decompiled code only, it has to happen.
-                    // Otherwise it will cause a desync, which is more important to prevent.
-                    if (File.Exists(path))
-                        File.Delete(path);
+                    mainWindow.ShowError("Error during writing of GML code to profile:\n" + exc);
                 }
-            }
-            catch (Exception exc)
-            {
-                mainWindow.ShowError("Error during writing of GML code to profile:\n" + exc);
             }
 
             // Invalidate gettext if necessary
@@ -1035,8 +1057,6 @@ namespace UndertaleModTool
 
             // Decompile new code
             await DecompileCode(code, false, dialog);
-
-            //GMLCacheChanged.Add() is inside DecompileCode()
         }
         private void DecompiledEditor_LostFocus(object sender, RoutedEventArgs e)
         {
@@ -1091,7 +1111,6 @@ namespace UndertaleModTool
             {
                 var instructions = Assembler.Assemble(DisassemblyEditor.Text, data);
                 code.Replace(instructions);
-                mainWindow.NukeProfileGML(code.Name.Content);
             }
             catch (Exception ex)
             {
@@ -1112,8 +1131,6 @@ namespace UndertaleModTool
 
             if (!DisassemblyEditor.IsReadOnly)
             {
-                data.GMLCacheChanged.Add(code.Name.Content);
-
                 if (mainWindow.IsSaving)
                 {
                     mainWindow.IsSaving = false;
@@ -1238,6 +1255,10 @@ namespace UndertaleModTool
                         ContextMenuDark contextMenu = new();
                         foreach (UndertaleObject obj in possibleObjects)
                         {
+                            if (obj is null)
+                            {
+                                continue;
+                            }
                             MenuItemDark item = new();
                             item.Header = obj.ToString().Replace("_", "__");
                             item.PreviewMouseDown += (sender2, ev2) =>
@@ -1427,7 +1448,7 @@ namespace UndertaleModTool
                             else
                             {
                                 // Resolve 2.3 sub-functions for their parent entry
-                                if (data.GlobalFunctions?.NameToFunction?.TryGetValue(nameText, out Underanalyzer.IGMFunction f) == true)
+                                if (data.GlobalFunctions?.TryGetFunction(nameText, out Underanalyzer.IGMFunction f) == true)
                                 {
                                     ScriptsDict.TryGetValue(f.Name.Content, out val);
                                     val = (val as UndertaleScript)?.Code?.ParentEntry;
@@ -1454,7 +1475,7 @@ namespace UndertaleModTool
                         if (val is UndertaleScript)
                             val = null; // in GMS2.3 scripts are never referenced directly
 
-                        if (data.GlobalFunctions?.NameToFunction?.TryGetValue(nameText, out Underanalyzer.IGMFunction globalFunc) == true &&
+                        if (data.GlobalFunctions?.TryGetFunction(nameText, out Underanalyzer.IGMFunction globalFunc) == true &&
                             globalFunc is UndertaleFunction utGlobalFunc)
                         {
                             // Try getting script that this function reference belongs to
