@@ -1,8 +1,6 @@
 ﻿using Microsoft.Win32;
 using System;
-using System.Drawing;
 using System.Windows;
-using System.IO;
 using UndertaleModLib.Models;
 using UndertaleModLib.Util;
 using System.Windows.Controls;
@@ -11,6 +9,9 @@ using System.Windows.Data;
 using UndertaleModTool.Windows;
 using UndertaleModLib;
 using WpfAnimatedGif;
+using ImageMagick;
+using System.Windows.Media.Imaging;
+using System.ComponentModel;
 
 namespace UndertaleModTool
 {
@@ -21,6 +22,16 @@ namespace UndertaleModTool
     {
         private static readonly MainWindow mainWindow = Application.Current.MainWindow as MainWindow;
 
+        /// <summary>
+        /// Handle on the texture page item we're listening for updates from.
+        /// </summary>
+        private UndertaleTexturePageItem _textureItemContext = null;
+
+        /// <summary>
+        /// Handle on the texture data where we're listening for updates from.
+        /// </summary>
+        private UndertaleEmbeddedTexture.TexData _textureDataContext = null;
+
         public UndertaleTexturePageItemEditor()
         {
             InitializeComponent();
@@ -30,6 +41,114 @@ namespace UndertaleModTool
             ((System.Windows.Controls.Image)mainWindow.FindName("FloweyBubble")).Opacity = 0;
 
             ((Label)this.FindName("TexutrePagesObjectLabel")).Content = ((Label)mainWindow.FindName("ObjectLabel")).Content;
+            DataContextChanged += SwitchDataContext;
+            Unloaded += UnloadTexture;
+        }
+
+        private void UpdateImages(UndertaleTexturePageItem item)
+        {
+            if (item.TexturePage?.TextureData?.Image is null)
+            {
+                ItemTextureBGImage.Source = null;
+                ItemTextureImage.Source = null;
+                return;
+            }
+
+            GMImage image = item.TexturePage.TextureData.Image;
+            BitmapSource bitmap = mainWindow.GetBitmapSourceForImage(image);
+            ItemTextureBGImage.Source = bitmap;
+            ItemTextureImage.Source = bitmap;
+        }
+
+        private void SwitchDataContext(object sender, DependencyPropertyChangedEventArgs e)
+        {
+            UndertaleTexturePageItem item = (DataContext as UndertaleTexturePageItem);
+            if (item is null)
+                return;
+
+            // Load current image
+            UpdateImages(item);
+
+            // Start listening for texture page updates
+            if (_textureItemContext is not null)
+            {
+                _textureItemContext.PropertyChanged -= ReloadTexturePage;
+            }
+            _textureItemContext = item;
+            _textureItemContext.PropertyChanged += ReloadTexturePage;
+
+            // Start listening for texture image updates
+            if (_textureDataContext is not null)
+            {
+                _textureDataContext.PropertyChanged -= ReloadTextureImage;
+            }
+
+            if (item.TexturePage?.TextureData is not null)
+            {
+                _textureDataContext = item.TexturePage.TextureData;
+                _textureDataContext.PropertyChanged += ReloadTextureImage;
+            }
+        }
+
+        private void ReloadTexturePage(object sender, PropertyChangedEventArgs e)
+        {
+            // Invoke dispatcher to only perform updates on UI thread
+            Dispatcher.Invoke(() =>
+            {
+                UndertaleTexturePageItem item = (DataContext as UndertaleTexturePageItem);
+                if (item is null)
+                    return;
+
+                if (e.PropertyName != nameof(UndertaleTexturePageItem.TexturePage))
+                    return;
+
+                UpdateImages(item);
+
+                // Start listening for (new) texture image updates
+                if (_textureDataContext is not null)
+                {
+                    _textureDataContext.PropertyChanged -= ReloadTextureImage;
+                }
+                _textureDataContext = item.TexturePage.TextureData;
+                _textureDataContext.PropertyChanged += ReloadTextureImage;
+            });
+        }
+
+        private void ReloadTextureImage(object sender, PropertyChangedEventArgs e)
+        {
+            // Invoke dispatcher to only perform updates on UI thread
+            Dispatcher.Invoke(() =>
+            {
+                UndertaleTexturePageItem item = (DataContext as UndertaleTexturePageItem);
+                if (item is null)
+                    return;
+
+                if (e.PropertyName != nameof(UndertaleEmbeddedTexture.TexData.Image))
+                    return;
+
+                // If the texture's image was updated, reload it
+                UpdateImages(item);
+            });
+        }
+
+        private void UnloadTexture(object sender, RoutedEventArgs e)
+        {
+            ItemTextureBGImage.Source = null;
+            ItemTextureImage.Source = null;
+
+            // Stop listening for texture page updates
+            if (_textureItemContext is not null)
+            {
+                _textureItemContext.PropertyChanged -= ReloadTexturePage;
+                _textureItemContext = null;
+            }
+
+            // Stop listening for texture image updates
+            if (_textureDataContext is not null)
+            {
+                _textureDataContext.PropertyChanged -= ReloadTextureImage;
+                _textureDataContext = null;
+            }
         }
 
         private void Import_Click(object sender, RoutedEventArgs e)
@@ -44,9 +163,9 @@ namespace UndertaleModTool
 
             try
             {
-                Bitmap image = TextureWorker.ReadImageFromFile(dlg.FileName);
-                image.SetResolution(96.0F, 96.0F);
-                (this.DataContext as UndertaleTexturePageItem).ReplaceTexture(image);
+                using MagickImage image = TextureWorker.ReadBGRAImageFromFile(dlg.FileName);
+                UndertaleTexturePageItem item = DataContext as UndertaleTexturePageItem;
+                item.ReplaceTexture(image);
 
                 // Refresh the image of "ItemDisplay"
                 if (ItemDisplay.FindName("RenderAreaBorder") is not Border border)
@@ -70,16 +189,15 @@ namespace UndertaleModTool
 
             if (dlg.ShowDialog() == true)
             {
-                TextureWorker worker = new TextureWorker();
+                using TextureWorker worker = new();
                 try
                 {
-                    worker.ExportAsPNG((UndertaleTexturePageItem)this.DataContext, dlg.FileName);
+                    worker.ExportAsPNG((UndertaleTexturePageItem)DataContext, dlg.FileName);
                 }
                 catch (Exception ex)
                 {
                     mainWindow.ShowError("Failed to export file: " + ex.Message, "Failed to export file");
                 }
-                worker.Cleanup();
             }
         }
 
@@ -97,7 +215,7 @@ namespace UndertaleModTool
             }
             catch (Exception ex)
             {
-                mainWindow.ShowError("An error occured in the object references related window.\n" +
+                mainWindow.ShowError("An error occurred in the object references related window.\n" +
                                      $"Please report this on GitHub.\n\n{ex}");
             }
             finally
