@@ -18,6 +18,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using UndertaleModLib.Models;
 using Newtonsoft.Json;
+using UndertaleModLib.Compiler;
 
 namespace UndertaleModCli;
 
@@ -160,7 +161,7 @@ public partial class Program : IScriptInterface
             dumpCommand,
             replaceCommand
         };
-        rootCommand.Description = "CLI tool for modding, decompiling and unpacking Undertale (and other Game Maker: Studio games)!";
+        rootCommand.Description = "CLI tool for modding, decompiling and unpacking Undertale (and other GameMaker games)!";
         Parser commandLine = new CommandLineBuilder(rootCommand)
             .UseDefaults() // automatically configures dotnet-suggest
             .Build();
@@ -182,7 +183,7 @@ public partial class Program : IScriptInterface
         this.ExePath = Environment.CurrentDirectory;
         this.Output = output;
 
-        this.Data = ReadDataFile(datafile, this.Verbose ? WarningHandler : DummyHandler, this.Verbose ? MessageHandler : DummyHandler);
+        this.Data = ReadDataFile(datafile, WarningHandler, this.Verbose ? MessageHandler : DummyHandler);
 
         FinishedMessageEnabled = true;
         this.CliScriptOptions = ScriptOptions.Default
@@ -195,6 +196,7 @@ public partial class Program : IScriptInterface
                 typeof(JsonConvert).GetTypeInfo().Assembly,
                 typeof(System.Text.RegularExpressions.Regex).GetTypeInfo().Assembly,
                 typeof(TextureWorker).GetTypeInfo().Assembly,
+                typeof(ImageMagick.MagickImage).GetTypeInfo().Assembly,
                 typeof(Underanalyzer.Decompiler.DecompileContext).Assembly)
             // "WithEmitDebugInformation(true)" not only lets us to see a script line number which threw an exception,
             // but also provides other useful debug info when we run UMT in "Debug".
@@ -584,7 +586,8 @@ public partial class Program : IScriptInterface
         if (!Data.IsYYC())
         {
             Console.WriteLine($"{Data.Code.Count} Code Entries, {Data.Variables.Count} Variables, {Data.Functions.Count} Functions");
-            Console.WriteLine($"{Data.CodeLocals.Count} Code locals, {Data.Strings.Count} Strings, {Data.EmbeddedTextures.Count} Embedded Textures");
+            var codeLocalsInfo = Data.CodeLocals is not null ? $"{Data.CodeLocals.Count} Code locals, " : "";
+            Console.WriteLine($"{codeLocalsInfo}{Data.Strings.Count} Strings, {Data.EmbeddedTextures.Count} Embedded Textures");
         }
         else
         {
@@ -658,7 +661,13 @@ public partial class Program : IScriptInterface
         {
             if (Verbose)
                 Console.WriteLine($"Dumping {texture.Name}");
-            File.WriteAllBytes($"{directory}/{texture.Name.Content}.png", texture.TextureData.TextureBlob);
+            if (texture.TextureData.Image is not GMImage image)
+            {
+                Console.WriteLine($"{texture.Name} has no image assigned, skipping");
+                continue;
+            }
+            using FileStream fs = new($"{directory}/{texture.Name.Content}.png", FileMode.Create);
+            texture.TextureData.Image.SavePng(fs);
         }
     }
 
@@ -669,18 +678,138 @@ public partial class Program : IScriptInterface
     /// <param name="fileToReplace">File path which should replace the code entry.</param>
     private void ReplaceCodeEntryWithFile(string codeEntry, FileInfo fileToReplace)
     {
-        UndertaleCode code = Data.Code.ByName(codeEntry);
-
-        if (code == null)
-        {
-            Console.Error.WriteLine($"Data file does not contain a code entry named {codeEntry}!");
-            return;
-        }
-
         if (Verbose)
             Console.WriteLine("Replacing " + codeEntry);
 
-        ImportGMLString(codeEntry, File.ReadAllText(fileToReplace.FullName));
+        // Read source code from file
+        string gmlCode = File.ReadAllText(fileToReplace.FullName);
+
+        // Link code to object events manually only if collision events are used
+        CompileResult result = CompileResult.UnsuccessfulResult;
+        bool manualLink = false;
+        const string objectPrefix = "gml_Object_";
+        if (codeEntry.StartsWith(objectPrefix, StringComparison.Ordinal))
+        {
+            // Parse object event. First, find positions of last two underscores in name.
+            int lastUnderscore = codeEntry.LastIndexOf('_');
+            int secondLastUnderscore = codeEntry.LastIndexOf('_', lastUnderscore - 1);
+            if (lastUnderscore <= 0 || secondLastUnderscore <= 0)
+            {
+                Console.Error.WriteLine($"Failed to parse object code entry name: \"{codeEntry}\"");
+                return;
+            }
+
+            // Extract object name, event type, and event subtype
+            ReadOnlySpan<char> objectName = codeEntry.AsSpan(new Range(objectPrefix.Length, secondLastUnderscore));
+            ReadOnlySpan<char> eventType = codeEntry.AsSpan(new Range(secondLastUnderscore + 1, lastUnderscore));
+            if (!uint.TryParse(codeEntry.AsSpan(lastUnderscore + 1), out uint eventSubtype))
+            {
+                // No number at the end of the name; parse it out as best as possible (may technically be ambiguous sometimes...).
+                // It should be a collision event, though.
+                manualLink = true;
+                ReadOnlySpan<char> nameAfterPrefix = codeEntry.AsSpan(objectPrefix.Length);
+                const string collisionSeparator = "_Collision_";
+                int collisionSeparatorPos = nameAfterPrefix.LastIndexOf(collisionSeparator);
+                if (collisionSeparatorPos != -1)
+                {
+                    // Split out the actual object name and the collision subtype
+                    objectName = nameAfterPrefix[0..collisionSeparatorPos];
+                    ReadOnlySpan<char> collisionSubtype = nameAfterPrefix[(collisionSeparatorPos + collisionSeparator.Length)..];
+
+                    if (Data.IsVersionAtLeast(2, 3))
+                    {
+                        // GameMaker 2.3+ uses the object name for the collision subtype
+                        int objectIndex = Data.GameObjects.IndexOfName(collisionSubtype);
+                        if (objectIndex >= 0)
+                        {
+                            // Object already exists; use its ID as a subtype
+                            eventSubtype = (uint)objectIndex;
+                        }
+                        else
+                        {
+                            // Need to create a new object
+                            eventSubtype = (uint)Data.GameObjects.Count;
+                            Data.GameObjects.Add(new()
+                            {
+                                Name = Data.Strings.MakeString(collisionSubtype.ToString())
+                            });
+                        }
+                    }
+                    else
+                    {
+                        // Pre-2.3 GMS2 versions use GUIDs... need to resolve it
+                        eventSubtype = ReduceCollisionValue(GetCollisionValueFromCodeNameGUID(codeEntry));
+                        ReassignGUIDs(collisionSubtype.ToString(), ReduceCollisionValue(GetCollisionValueFromCodeNameGUID(codeEntry)));
+                    }
+                }
+                else
+                {
+                    Console.Error.WriteLine($"Failed to parse event type and subtype for \"{codeEntry}\".");
+                    return;
+                }
+            }
+            else if (eventType.SequenceEqual("Collision"))
+            {
+                // Handle collision events with object ID at the end of the name
+                manualLink = true;
+                if (eventSubtype >= Data.GameObjects.Count)
+                {
+                    if (ScriptQuestion($"Object of ID {eventSubtype} was not found.\nAdd new object? (will be ID {Data.GameObjects.Count})"))
+                    {
+                        // Create new object at end of game object list
+                        eventSubtype = (uint)Data.GameObjects.Count;
+                        Data.GameObjects.Add(new()
+                        {
+                            Name = Data.Strings.MakeString(
+                                SimpleTextInput("Enter object name", $"Enter object name for ID {eventSubtype}", "", false))
+                        });
+                    }
+                    else
+                    {
+                        // It *needs* to have a valid value, make the user specify one
+                        eventSubtype = ReduceCollisionValue([uint.MaxValue]);
+                    }
+                }
+            }
+
+            // If manually linking, do so
+            if (manualLink)
+            {
+                // Create new object if necessary
+                UndertaleGameObject obj = Data.GameObjects.ByName(objectName);
+                if (obj is null)
+                {
+                    obj = new()
+                    {
+                        Name = Data.Strings.MakeString(objectName.ToString())
+                    };
+                    Data.GameObjects.Add(obj);
+                }
+
+                // Link to object's event with a blank code entry
+                UndertaleCode manualCode = UndertaleCode.CreateEmptyEntry(Data, codeEntry);
+                CodeImportGroup.LinkEvent(obj, manualCode, EventType.Collision, eventSubtype);
+
+                // Perform code import using manual code entry
+                CodeImportGroup group = new(Data);
+                group.QueueReplace(manualCode, gmlCode);
+                result = group.Import();
+            }
+        }
+
+        // When not manually linking, just let a code import group do it during importing
+        if (!manualLink)
+        {
+            CodeImportGroup group = new(Data);
+            group.QueueReplace(codeEntry, gmlCode);
+            result = group.Import();
+        }
+
+        // Error if import failed
+        if (!result.Successful)
+        {
+            Console.Error.WriteLine("Code import unsuccessful:\n" + result.PrintAllErrors(false));
+        }
     }
 
     /// <summary>
@@ -701,7 +830,7 @@ public partial class Program : IScriptInterface
         if (Verbose)
             Console.WriteLine("Replacing " + textureEntry);
 
-        texture.TextureData.TextureBlob = File.ReadAllBytes(fileToReplace.FullName);
+        texture.TextureData.Image = GMImage.FromPng(File.ReadAllBytes(fileToReplace.FullName));
     }
 
     /// <summary>
@@ -713,7 +842,7 @@ public partial class Program : IScriptInterface
         string lines;
         try
         {
-            lines = File.ReadAllText(path);
+            lines = File.ReadAllText(path, Encoding.UTF8);
         }
         catch (Exception exc)
         {
@@ -731,8 +860,8 @@ public partial class Program : IScriptInterface
     /// Evaluates and executes given C# code.
     /// </summary>
     /// <param name="code">The C# string to execute</param>
-    /// <param name="scriptFile">The path to the script file where <see cref="code"/> was executed from.
-    /// Leave as null, if it wasn't executed from a script file</param>
+    /// <param name="scriptFile">The path to the script file where <paramref name="code"/> was loaded from.
+    /// Leave as null, if it wasn't executed from a script file.</param>
     private void RunCSharpCode(string code, string scriptFile = null)
     {
         if (Verbose)
@@ -740,7 +869,7 @@ public partial class Program : IScriptInterface
 
         try
         {
-            CSharpScript.EvaluateAsync(code, CliScriptOptions, this, typeof(IScriptInterface)).GetAwaiter().GetResult();
+            CSharpScript.EvaluateAsync(code, CliScriptOptions.WithFilePath(scriptFile ?? "").WithFileEncoding(Encoding.UTF8), this, typeof(IScriptInterface)).GetAwaiter().GetResult();
             ScriptExecutionSuccess = true;
             ScriptErrorMessage = "";
         }
@@ -768,23 +897,41 @@ public partial class Program : IScriptInterface
     /// Saves the currently loaded <see cref="Data"/> to an output path.
     /// </summary>
     /// <param name="outputPath">The path where to save the data.</param>
+    /// <exception cref="IOException">If saving fails</exception>
     private void SaveDataFile(string outputPath)
     {
         if (Verbose)
             Console.WriteLine($"Saving new data file to '{outputPath}'");
+        try 
+        {
+            // Save data.win to temp file
+            using (FileStream fs = new(outputPath + "temp", FileMode.Create, FileAccess.Write))
+            {
+                UndertaleIO.Write(fs, Data, MessageHandler);
+            }
 
-        using FileStream fs = new FileInfo(outputPath).OpenWrite();
-        UndertaleIO.Write(fs, Data, MessageHandler);
-        if (Verbose)
-            Console.WriteLine($"Saved data file to '{outputPath}'");
+            // If we're executing this, the saving was successful. So we can replace the new temp file
+            // with the older file, if it exists.
+            File.Move(outputPath + "temp", outputPath, true);
+
+            if (Verbose)
+                Console.WriteLine($"Saved data file to '{outputPath}'");
+        }
+        catch (Exception e) 
+        {
+            // Delete the temporary file in case we partially wrote it
+            if (File.Exists(outputPath + "temp"))
+                File.Delete(outputPath + "temp");
+            throw new IOException($"Could not save data file: {e.Message}");
+        }
     }
 
     /// <summary>
     /// Read supplied filename and return the data file.
     /// </summary>
     /// <param name="datafile">The datafile to read</param>
-    /// <param name="warningHandler">Handler for Warnings</param>
-    /// <param name="messageHandler">Handler for Messages</param>
+    /// <param name="warningHandler">Handler for warnings</param>
+    /// <param name="messageHandler">Handler for messages</param>
     /// <returns></returns>
     /// <exception cref="FileNotFoundException">If the data file cannot be found</exception>
     private static UndertaleData ReadDataFile(FileInfo datafile, WarningHandlerDelegate warningHandler = null, MessageHandlerDelegate messageHandler = null)
@@ -829,7 +976,8 @@ public partial class Program : IScriptInterface
     /// A simple warning handler that prints warnings to console.
     /// </summary>
     /// <param name="warning">The warning to print</param>
-    private static void WarningHandler(string warning) => Console.WriteLine($"[WARNING]: {warning}");
+    /// <param name="isImportant">Whether the warning is important (may lead to data corruption)</param>
+    private static void WarningHandler(string warning, bool isImportant) => Console.WriteLine($"[WARNING]: {warning}");
 
     /// <summary>
     /// A simple message handler that prints messages to console.
