@@ -11,7 +11,7 @@ namespace UndertaleModLib.Models;
 /// A function entry as it's used in a GameMaker data file.
 /// </summary>
 [PropertyChanged.AddINotifyPropertyChangedInterface]
-public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.ReferencedObject, IStaticChildObjectsSize, IDisposable, IGMFunction
+public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.IReferencedObject, IStaticChildObjectsSize, IDisposable, IGMFunction
 {
     /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
     public static readonly uint ChildObjectsSize = 12;
@@ -32,16 +32,11 @@ public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.Re
     /// How often this <see cref="UndertaleFunction"/> is referenced in code.
     /// </summary>
     public uint Occurrences { get; set; }
-
+    
     /// <summary>
     /// The first instruction in this function.
     /// </summary>
     public UndertaleInstruction FirstAddress { get; set; }
-
-    /// <summary>
-    /// TODO: some kind of helper variable for something in gms2.3
-    /// </summary>
-    public bool Autogenerated { get; set; } = false; // Whether a gml_Script_... kind of function in 2.3
 
     [Obsolete("This variable has been renamed to NameStringID.")]
     public int UnknownChainEndingValue { get => NameStringID; set => NameStringID = value; }
@@ -74,7 +69,7 @@ public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.Re
                 FirstAddress = reader.GetUndertaleObjectAtAddress<UndertaleInstruction>(reader.ReadUInt32() - 4);
             else
                 FirstAddress = reader.ReadUndertaleObjectPointer<UndertaleInstruction>();
-            UndertaleInstruction.Reference<UndertaleFunction>.ParseReferenceChain(reader, this);
+            UndertaleInstruction.ParseReferenceChain(reader, this);
         }
         else
         {
@@ -112,11 +107,32 @@ public class UndertaleCodeLocals : UndertaleNamedResource, IDisposable
     /// The name of the code local entry which corresponds to a code entry.
     /// </summary>
     public UndertaleString Name { get; set; }
-
+    
     /// <summary>
     /// A collection of local variables.
     /// </summary>
     public ObservableCollection<LocalVar> Locals { get; private set; } = new ObservableCollection<LocalVar>();
+
+    /// <summary>
+    /// Creates an empty code locals entry with the given name.
+    /// </summary>
+    /// <param name="data">Data to add the new code to.</param>
+    /// <param name="name">Name of the new code locals entry to create.</param>
+    /// <returns>The new code locals entry.</returns>
+    public static UndertaleCodeLocals CreateEmptyEntry(UndertaleData data, UndertaleString name)
+    {
+        UndertaleCodeLocals locals = new()
+        {
+            Name = name
+        };
+        locals.Locals.Add(new()
+        {
+            Name = data.Strings.MakeString("arguments", out int argumentsStringId),
+            Index = data.IsVersionAtLeast(2, 3) ? (uint)argumentsStringId : 0
+        });
+        data.CodeLocals.Add(locals);
+        return locals;
+    }
 
     /// <inheritdoc />
     public void Serialize(UndertaleWriter writer)
@@ -174,7 +190,7 @@ public class UndertaleCodeLocals : UndertaleNamedResource, IDisposable
         /// TODO: have no idea what this index does.
         /// </summary>
         public uint Index { get; set; }
-
+        
         /// <summary>
         /// The name of the local variable.
         /// </summary>
