@@ -305,12 +305,12 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
     public List<long> GMS2RandomUID { get; set; } = new List<long>();
 
     /// <summary>
-    /// The FPS of the data file. Game Maker Studio: 2 only.
+    /// The FPS of the data file. GameMaker Studio 2 only.
     /// </summary>
     public float GMS2FPS { get; set; } = 30.0f;
 
     /// <summary>
-    /// Whether the data file allows statistics. Game Maker Studio: 2 only.
+    /// Whether the data file allows statistics. GameMaker Studio 2 only.
     /// </summary>
     public bool GMS2AllowStatistics { get; set; } = true;
 
@@ -330,7 +330,9 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
         (uint Major, uint Minor, uint Release, uint Build, BranchType Branch) detectedVer = readVersion;
 
         // Some GMS2+ version detection. The rest is spread around, mostly in UndertaleChunks.cs
-        if (reader.AllChunkNames.Contains("PSEM"))      // 2023.2, not present on LTS
+        if (reader.AllChunkNames.Contains("UILR"))      // 2024.13, not present on LTS
+            detectedVer = (2024, 13, 0, 0, BranchType.Post2022_0);
+        else if (reader.AllChunkNames.Contains("PSEM")) // 2023.2, not present on LTS
             detectedVer = (2023, 2, 0, 0, BranchType.Post2022_0);
         else if (reader.AllChunkNames.Contains("FEAT")) // 2022.8
             detectedVer = (2022, 8, 0, 0, BranchType.Pre2022_0);
@@ -340,15 +342,6 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
             detectedVer = (2, 3, 0, 0, BranchType.Pre2022_0);
         else if (reader.AllChunkNames.Contains("TGIN")) // 2.2.1
             detectedVer = (2, 2, 1, 0, BranchType.Pre2022_0);
-
-        if (detectedVer.Major > 2 || (detectedVer.Major == 2 && detectedVer.Minor >= 3))
-        {
-            CompileContext.GMS2_3 = true;
-        }
-        else
-        {
-            CompileContext.GMS2_3 = false;
-        }
 
         return detectedVer;
     }
@@ -479,24 +472,13 @@ public class UndertaleGeneralInfo : UndertaleObject, IDisposable
         if (reader.ReadOnlyGEN8)
             return;
 
-        var detectedVer = TestForCommonGMSVersions(reader, (Major, Minor, Release, Build, Branch));
-        (Major, Minor, Release, Build, Branch) = detectedVer;
-
-        if (reader.undertaleData.GeneralInfo is not null)
-        {
-            var prevGenInfo = reader.undertaleData.GeneralInfo;
-            // If previous version is greater than current
-            if (prevGenInfo.Major > Major
-                || prevGenInfo.Major == Major && prevGenInfo.Minor > Minor
-                || prevGenInfo.Major == Major && prevGenInfo.Minor == Minor && prevGenInfo.Release > Release
-                || prevGenInfo.Major == Major && prevGenInfo.Minor == Minor && prevGenInfo.Release == Release && prevGenInfo.Build > Build)
-            {
-                Major = prevGenInfo.Major;
-                Minor = prevGenInfo.Minor;
-                Release = prevGenInfo.Release;
-                Build = prevGenInfo.Build;
-            }
-        }
+        // TestForCommonGMSVersions is run during the object counting phase, so the previous general info is always accurate.
+        var prevGenInfo = reader.undertaleData.GeneralInfo;
+        Major = prevGenInfo.Major;
+        Minor = prevGenInfo.Minor;
+        Release = prevGenInfo.Release;
+        Build = prevGenInfo.Build;
+        Branch = prevGenInfo.Branch;
 
         DefaultWindowWidth = reader.ReadUInt32();
         DefaultWindowHeight = reader.ReadUInt32();
@@ -712,7 +694,13 @@ public class UndertaleOptions : UndertaleObject, IDisposable
         UseRearTouch = 0x2000000,
         UseFastCollision = 0x4000000,
         FastCollisionCompatibility = 0x8000000,
-        DisableSandbox = 0x10000000
+        DisableSandbox = 0x10000000,
+        EnableCopyOnWrite = 0x20000000,
+        LegacyJsonParsing = 0x40000000,
+        LegacyNumberConversion = 0x80000000,
+        LegacyOtherBehavior = 0x100000000,
+        AudioErrorBehavior = 0x200000000,
+        AllowInstanceChange = 0x400000000
     }
 
     /// <summary>
