@@ -7,8 +7,6 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Windows;
-using UndertaleModLib.Models;
 using UndertaleModLib.Util;
 
 namespace UndertaleModLib.Models;
@@ -37,13 +35,17 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         /// </summary>
         DoNotClearDisplayBuffer = 4,
         /// <summary>
-        /// Whether the room was made in Game Maker: Studio 2.
+        /// Whether the room was made in GameMaker Studio 2 or above.
         /// </summary>
         IsGMS2 = 131072,
         /// <summary>
-        /// Whether the room was made in Game Maker: Studio 2.3.
+        /// Whether the room was made in GameMaker Studio 2.3 or above.
         /// </summary>
-        IsGMS2_3 = 65536
+        IsGMS2_3 = 65536,
+        /// <summary>
+        /// Whether the room was made in GameMaker 2024.13 or above.
+        /// </summary>
+        IsGM2024_13 = 262144
     }
 
     /// <summary>
@@ -161,7 +163,12 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
     public UndertalePointerList<Tile> Tiles { get; set; } = new UndertalePointerList<Tile>();
 
     /// <summary>
-    /// The list of layers this room uses. Used in Game Maker Studio: 2 only, as <see cref="Backgrounds"/> and <see cref="Tiles"/> are empty there.
+    /// List of instance creation order IDs, used for the first room in the room order only, in GameMaker 2024.13 and above.
+    /// </summary>
+    public InstanceIDList InstanceCreationOrderIDs { get; set; } = null;
+
+    /// <summary>
+    /// The list of layers this room uses. Used in GameMaker Studio 2 only, as <see cref="Backgrounds"/> and <see cref="Tiles"/> are empty there.
     /// </summary>
     public UndertalePointerList<Layer> Layers { get => _layers; set { _layers = value; UpdateBGColorLayer(); OnPropertyChanged(); } }
 
@@ -276,20 +283,6 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
     /// <inheritdoc />
     public void Serialize(UndertaleWriter writer)
     {
-        if (writer.undertaleData.IsGameMaker2())
-        {
-            foreach (var layer in Layers)
-            {
-                if (layer.InstancesData != null)
-                {
-                    foreach (var inst in layer.InstancesData.Instances)
-                    {
-                        if (!GameObjects.Contains(inst))
-                            throw new Exception("Nonexistent instance " + inst.InstanceID);
-                    }
-                }
-            }
-        }
         writer.WriteUndertaleString(Name);
         writer.WriteUndertaleString(Caption);
         writer.Write(Width);
@@ -304,6 +297,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         writer.WriteUndertaleObjectPointer(Views);
         writer.WriteUndertaleObjectPointer(GameObjects);
         writer.WriteUndertaleObjectPointer(Tiles);
+        if (writer.undertaleData.IsVersionAtLeast(2024, 13))
+        {
+            writer.WriteUndertaleObjectPointer(InstanceCreationOrderIDs);
+        }
         writer.Write(World);
         writer.Write(Top);
         writer.Write(Left);
@@ -324,6 +321,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         writer.WriteUndertaleObject(Views);
         writer.WriteUndertaleObject(GameObjects);
         writer.WriteUndertaleObject(Tiles);
+        if (writer.undertaleData.IsVersionAtLeast(2024, 13))
+        {
+            writer.WriteUndertaleObject(InstanceCreationOrderIDs);
+        }
         if (writer.undertaleData.IsGameMaker2())
         {
             writer.WriteUndertaleObject(Layers);
@@ -350,6 +351,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         Views = reader.ReadUndertaleObjectPointer<UndertalePointerList<View>>();
         GameObjects = reader.ReadUndertaleObjectPointer<UndertalePointerList<GameObject>>();
         Tiles = reader.ReadUndertaleObjectPointer<UndertalePointerList<Tile>>();
+        if (reader.undertaleData.IsVersionAtLeast(2024, 13))
+        {
+            InstanceCreationOrderIDs = reader.ReadUndertaleObjectPointer<InstanceIDList>();
+        }
         World = reader.ReadBoolean();
         Top = reader.ReadUInt32();
         Left = reader.ReadUInt32();
@@ -370,6 +375,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         reader.ReadUndertaleObject(Views);
         reader.ReadUndertaleObject(GameObjects);
         reader.ReadUndertaleObject(Tiles);
+        if (reader.undertaleData.IsVersionAtLeast(2024, 13))
+        {
+            reader.ReadUndertaleObject(InstanceCreationOrderIDs);
+        }
         if (reader.undertaleData.IsGameMaker2())
         {
             reader.ReadUndertaleObject(Layers);
@@ -401,20 +410,12 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
                             layer.InstancesData.Instances.Add(gameObj);
                         else
                         {
-                            /* Attempt to resolve null objects.
-                             * Sometimes, the instance ID in GameObjects will end up a duplicate
-                             * of a previous ID, rather than the correct one.
-                             * So, we traverse the object list a little to find the correct one.
-                             * If you can get two broken objects in a row... it'll probably crash.
-                             */
-                            int foundIndex = GameObjects.IndexOf(GameObjects.ByInstanceID(id - 1));
-                            if (GameObjects.Count - 1 <= foundIndex)
+                            // Object with this instance ID doesn't actually exist. Create a nonexistent object in its place.
+                            layer.InstancesData.Instances.Add(new GameObject()
                             {
-                                Debug.WriteLine($"The object instance with ID {id} of a layer (ID {layer.LayerId}) is not found.");
-                                continue;
-                            }
-
-                            layer.InstancesData.Instances.Add(GameObjects[foundIndex + 1]);
+                                InstanceID = id,
+                                Nonexistent = true
+                            });
                         }
                     }
                 }
@@ -439,6 +440,13 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         uint tilesPtr = reader.ReadUInt32();
         uint layersPtr = 0;
         uint sequencesPtr = 0;
+
+        if (reader.undertaleData.IsVersionAtLeast(2024, 13))
+        {
+            // InstanceCreationOrderIDs
+            reader.Position += 4;
+            count += 1;
+        }
 
         reader.Position += 32;
 
@@ -504,11 +512,16 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             {
                 if (layer.LayerType == LayerType.Assets)
                     tileList = tileList.Concat(layer.AssetsData.LegacyTiles);
-                else if (layer.LayerType == LayerType.Tiles && layer.TilesData.TileData.Length != 0)
+                else if (layer.LayerType == LayerType.Tiles && layer.TilesData.TileData.Length != 0 && layer.TilesData.Background is not null)
                 {
-                    int w = (int)(Width / layer.TilesData.TilesX);
-                    int h = (int)(Height / layer.TilesData.TilesY);
-                    tileSizes[new(w, h)] = layer.TilesData.TilesX * layer.TilesData.TilesY;
+                    int w = (int)layer.TilesData.Background.GMS2TileWidth;
+                    int h = (int)layer.TilesData.Background.GMS2TileHeight;
+                    Point scale = new(w, h);
+                    uint numTiles = layer.TilesData.TilesX * layer.TilesData.TilesY;
+                    if (tileSizes.ContainsKey(scale))
+                        tileSizes[scale] += numTiles;
+                    else
+                        tileSizes[scale] = numTiles;
                 }
             }
 
@@ -556,10 +569,11 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         GC.SuppressFinalize(this);
 
         _creationCodeId.Dispose();
-        if (Flags.HasFlag(RoomEntryFlags.IsGMS2))
+        if (Flags.HasFlag(RoomEntryFlags.IsGMS2) || Flags.HasFlag(RoomEntryFlags.IsGM2024_13))
         {
             foreach (Layer layer in _layers)
                 layer?.Dispose();
+            InstanceCreationOrderIDs = null;
             _layers = null;
             Sequences = new();
         }
@@ -919,15 +933,13 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
     public class GameObject : UndertaleObject, IRoomObject, INotifyPropertyChanged, IDisposable,
                               IStaticChildObjCount, IStaticChildObjectsSize
     {
-        /// <inheritdoc cref="IStaticChildObjCount.ChildObjectCount" />
-        public static readonly uint ChildObjectCount = 2;
-
         /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
         public static readonly uint ChildObjectsSize = 36;
 
-        private UndertaleResourceById<UndertaleGameObject, UndertaleChunkOBJT> _objectDefinition = new();
-        private UndertaleResourceById<UndertaleCode, UndertaleChunkCODE> _creationCode = new();
-        private UndertaleResourceById<UndertaleCode, UndertaleChunkCODE> _preCreateCode = new();
+        private readonly UndertaleResourceById<UndertaleGameObject, UndertaleChunkOBJT> _objectDefinition = new();
+        private readonly UndertaleResourceById<UndertaleCode, UndertaleChunkCODE> _creationCode = new();
+        private readonly UndertaleResourceById<UndertaleCode, UndertaleChunkCODE> _preCreateCode = new();
+        private FloatAsInt _rotation;
 
         /// <summary>
         /// The x coordinate of this object.
@@ -970,7 +982,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         /// <summary>
         /// The rotation of this object.
         /// </summary>
-        public float Rotation { get; set; }
+        public float Rotation { get => _rotation.AsFloat(); set => _rotation = (FloatAsInt)value; }
 
         /// <summary>
         /// The pre creation code of this object.
@@ -986,6 +998,11 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         /// The image index of this object. Game Maker: Studio 2 only.
         /// </summary>
         public int ImageIndex { get; set; }
+
+        /// <summary>
+        /// Whether this game object actually exists or not; <see langword="true"/> if it does not exist.
+        /// </summary>
+        public bool Nonexistent { get; set; }
 
         /// <summary>
         /// A wrapper for <see cref="ImageIndex"/> that returns the value being wrapped around available frames of the sprite.<br/>
@@ -1087,13 +1104,13 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         public int YOffset => Y + SpriteYOffset;
 
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public virtual void Serialize(UndertaleWriter writer)
         {
             writer.Write(X);
             writer.Write(Y);
-            writer.WriteUndertaleObject(_objectDefinition);
+            _objectDefinition.Serialize(writer);
             writer.Write(InstanceID);
-            writer.WriteUndertaleObject(_creationCode);
+            _creationCode.Serialize(writer);
             writer.Write(ScaleX);
             writer.Write(ScaleY);
             if (writer.undertaleData.IsVersionAtLeast(2, 2, 2, 302))
@@ -1102,19 +1119,19 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
                 writer.Write(ImageIndex);
             }
             writer.Write(Color);
-            writer.Write(Rotation);
+            writer.Write(_rotation.AsUInt());
             if (writer.undertaleData.GeneralInfo.BytecodeVersion >= 16) // TODO: is that dependent on bytecode or something else?
-                writer.WriteUndertaleObject(_preCreateCode);         // Note: Appears in GM:S 1.4.9999 as well, so that's probably the closest it gets
+                _preCreateCode.Serialize(writer);                       // Note: Appears in GM:S 1.4.9999 as well, so that's probably the closest it gets
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public virtual void Unserialize(UndertaleReader reader)
         {
             X = reader.ReadInt32();
             Y = reader.ReadInt32();
-            _objectDefinition = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleGameObject, UndertaleChunkOBJT>>();
+            _objectDefinition.Unserialize(reader);
             InstanceID = reader.ReadUInt32();
-            _creationCode = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleCode, UndertaleChunkCODE>>();
+            _creationCode.Unserialize(reader);
             ScaleX = reader.ReadSingle();
             ScaleY = reader.ReadSingle();
             if (reader.undertaleData.IsVersionAtLeast(2, 2, 2, 302))
@@ -1123,14 +1140,22 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
                 ImageIndex = reader.ReadInt32();
             }
             Color = reader.ReadUInt32();
-            Rotation = reader.ReadSingle();
-            if (reader.undertaleData.GeneralInfo.BytecodeVersion >= 16) // TODO: is that dependent on bytecode or something else?
-                _preCreateCode = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleCode, UndertaleChunkCODE>>(); // Note: Appears in GM:S 1.4.9999 as well, so that's probably the closest it gets
+            _rotation = new FloatAsInt(reader.ReadUInt32());
+            if (reader.undertaleData.GeneralInfo.BytecodeVersion >= 16)
+                _preCreateCode.Unserialize(reader); // Note: Appears in GM:S 1.4.9999 as well, so that's probably the closest it gets
         }
 
         public override string ToString()
         {
-            return "Instance " + InstanceID + " of " + (ObjectDefinition?.Name?.Content ?? "?");
+            if (Nonexistent)
+            {
+                return $"Instance {InstanceID} (nonexistent)";
+            }
+            if (ObjectDefinition is null)
+            {
+                return $"Instance {InstanceID} (empty instance)";
+            }
+            return $"Instance {InstanceID} of {ObjectDefinition.Name?.Content ?? "(null object name)"}";
         }
 
         /// <inheritdoc/>
@@ -1157,7 +1182,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         public static readonly uint ChildObjectsSize = 48;
 
         /// <summary>
-        /// Whether this tile is from an asset layer.<br/>
+        /// Whether this tile represents a <see cref="UndertaleSprite"/> or a <see cref="UndertaleBackground"/>.<br/>
         /// <see langword="true"/> for GameMaker Studio: 2 games, otherwise <see langword="false"/>.
         /// </summary>
         /// <remarks>
@@ -1209,8 +1234,8 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
 
         /// <summary>
         /// From which object this tile stems from.<br/>
-        /// Will return a <see cref="UndertaleBackground"/> if <see cref="spriteMode"/> is <see langword="true"/>,
-        /// a <see cref="UndertaleSprite"/> if it's <see langword="false"/>.
+        /// Will return a <see cref="UndertaleSprite"/> if <see cref="spriteMode"/> is <see langword="true"/>,
+        /// a <see cref="UndertaleBackground"/> if it's <see langword="false"/>.
         /// </summary>
         public UndertaleNamedResource ObjectDefinition
         {
@@ -1225,12 +1250,12 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         /// <summary>
         /// The x coordinate of the tile in <see cref="ObjectDefinition"/>.
         /// </summary>
-        public uint SourceX { get; set; }
+        public int SourceX { get; set; }
 
         /// <summary>
         /// The y coordinate of the tile in <see cref="ObjectDefinition"/>.
         /// </summary>
-        public uint SourceY { get; set; }
+        public int SourceY { get; set; }
 
         /// <summary>
         /// The width of the tile.
@@ -1312,8 +1337,8 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
                 _spriteDefinition = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleSprite, UndertaleChunkSPRT>>();
             else
                 _backgroundDefinition = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleBackground, UndertaleChunkBGND>>();
-            SourceX = reader.ReadUInt32();
-            SourceY = reader.ReadUInt32();
+            SourceX = reader.ReadInt32();
+            SourceY = reader.ReadInt32();
             Width = reader.ReadUInt32();
             Height = reader.ReadUInt32();
             TileDepth = reader.ReadInt32();
@@ -1372,7 +1397,11 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         /// <summary>
         /// The layer is an effects layer.
         /// </summary>
-        Effect = 6
+        Effect = 6,
+        /// <summary>
+        /// The layer is a path layer, in 2024.13+ only.
+        /// </summary>
+        Path2 = 7
     }
 
     /// <summary>
@@ -1469,6 +1498,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             switch (LayerType)
             {
                 case LayerType.Path: break;
+                case LayerType.Path2: break;
                 case LayerType.Instances: writer.WriteUndertaleObject(InstancesData); break;
                 case LayerType.Tiles: writer.WriteUndertaleObject(TilesData); break;
                 case LayerType.Background: writer.WriteUndertaleObject(BackgroundData); break;
@@ -1500,7 +1530,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
 
             Data = LayerType switch
             {
-                LayerType.Path => null,
+                LayerType.Path or LayerType.Path2 => null,
                 LayerType.Instances => reader.ReadUndertaleObject<LayerInstancesData>(),
                 LayerType.Tiles => reader.ReadUndertaleObject<LayerTilesData>(),
                 LayerType.Background => reader.ReadUndertaleObject<LayerBackgroundData>(),
@@ -1651,6 +1681,14 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
                 }
             }
             public uint[][] TileData { get => _tileData; set { _tileData = value; OnPropertyChanged(); } }
+
+            /// <summary>
+            /// Invokes PropertyChanged on the TileData property. Use when mutating the TileData e.g alongside XAML.
+            /// </summary>
+            public void TileDataUpdated()
+            {
+                OnPropertyChanged(nameof(TileData));
+            }
 
             /// <inheritdoc />
             public event PropertyChangedEventHandler PropertyChanged;
@@ -1827,7 +1865,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
                     tile = reader.ReadUInt32();
 
                     // sanity check: run of 2 empty tiles
-                    if (length != 0x81)
+                    if (length != 0x81) 
                         throw new IOException("Expected 0x81, got 0x" + length.ToString("X2"));
                     if (tile != unchecked((uint)-1))
                         throw new IOException("Expected -1, got " + tile + " (0x" + tile.ToString("X8") + ")");
@@ -2298,9 +2336,6 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
 
     public class SpriteInstance : UndertaleObject, INotifyPropertyChanged, IStaticChildObjCount, IStaticChildObjectsSize, IDisposable
     {
-        /// <inheritdoc cref="IStaticChildObjCount.ChildObjectCount" />
-        public static readonly uint ChildObjectCount = 1;
-
         /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
         public static readonly uint ChildObjectsSize = 44;
 
@@ -2386,10 +2421,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         }
 
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public virtual void Serialize(UndertaleWriter writer)
         {
             writer.WriteUndertaleString(Name);
-            writer.WriteUndertaleObject(_sprite);
+            _sprite.Serialize(writer);
             writer.Write(X);
             writer.Write(Y);
             writer.Write(ScaleX);
@@ -2402,10 +2437,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public virtual void Unserialize(UndertaleReader reader)
         {
             Name = reader.ReadUndertaleString();
-            _sprite = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleSprite, UndertaleChunkSPRT>>();
+            _sprite.Unserialize(reader);
             X = reader.ReadInt32();
             Y = reader.ReadInt32();
             ScaleX = reader.ReadSingle();
@@ -2445,9 +2480,6 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
 
     public class SequenceInstance : UndertaleObject, INotifyPropertyChanged, IStaticChildObjCount, IStaticChildObjectsSize, IDisposable
     {
-        /// <inheritdoc cref="IStaticChildObjCount.ChildObjectCount" />
-        public static readonly uint ChildObjectCount = 1;
-
         /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
         public static readonly uint ChildObjectsSize = 44;
 
@@ -2468,10 +2500,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         public event PropertyChangedEventHandler PropertyChanged;
 
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public virtual void Serialize(UndertaleWriter writer)
         {
             writer.WriteUndertaleString(Name);
-            writer.WriteUndertaleObject(_sequence);
+            _sequence.Serialize(writer);
             writer.Write(X);
             writer.Write(Y);
             writer.Write(ScaleX);
@@ -2484,10 +2516,10 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public virtual void Unserialize(UndertaleReader reader)
         {
             Name = reader.ReadUndertaleString();
-            _sequence = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleSequence, UndertaleChunkSEQN>>();
+            _sequence.Unserialize(reader);
             X = reader.ReadInt32();
             Y = reader.ReadInt32();
             ScaleX = reader.ReadSingle();
@@ -2517,9 +2549,6 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
 
     public class ParticleSystemInstance : UndertaleObject, INotifyPropertyChanged, IStaticChildObjCount, IStaticChildObjectsSize, IDisposable
     {
-        /// <inheritdoc cref="IStaticChildObjCount.ChildObjectCount" />
-        public static readonly uint ChildObjectCount = 1;
-
         /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
         public static readonly uint ChildObjectsSize = 32;
 
@@ -2560,7 +2589,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         public void Serialize(UndertaleWriter writer)
         {
             writer.WriteUndertaleString(Name);
-            writer.WriteUndertaleObject(_particleSys);
+            _particleSys.Serialize(writer);
             writer.Write(X);
             writer.Write(Y);
             writer.Write(ScaleX);
@@ -2573,7 +2602,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         public void Unserialize(UndertaleReader reader)
         {
             Name = reader.ReadUndertaleString();
-            _particleSys = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleParticleSystem, UndertaleChunkPSYS>>();
+            _particleSys.Unserialize(reader);
             X = reader.ReadInt32();
             Y = reader.ReadInt32();
             ScaleX = reader.ReadSingle();
@@ -2608,9 +2637,6 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
 
     public class TextItemInstance : UndertaleObject, INotifyPropertyChanged, IStaticChildObjCount, IStaticChildObjectsSize, IDisposable
     {
-        /// <inheritdoc cref="IStaticChildObjCount.ChildObjectCount" />
-        public static readonly uint ChildObjectCount = 1;
-
         /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
         public static readonly uint ChildObjectsSize = 68;
 
@@ -2650,12 +2676,12 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         public bool Wrap { get; set; }
 
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public virtual void Serialize(UndertaleWriter writer)
         {
             writer.WriteUndertaleString(Name);
             writer.Write(X);
             writer.Write(Y);
-            writer.WriteUndertaleObject(_font);
+            _font.Serialize(writer);
             writer.Write(ScaleX);
             writer.Write(ScaleY);
             writer.Write(Rotation);
@@ -2672,12 +2698,12 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public virtual void Unserialize(UndertaleReader reader)
         {
             Name = reader.ReadUndertaleString();
             X = reader.ReadInt32();
             Y = reader.ReadInt32();
-            _font = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleFont, UndertaleChunkFONT>>();
+            _font.Unserialize(reader);
             ScaleX = reader.ReadSingle();
             ScaleY = reader.ReadSingle();
             Rotation = reader.ReadSingle();
@@ -2715,6 +2741,33 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             _font.Dispose();
             Name = null;
             Text = null;
+        }
+    }
+
+    /// <summary>
+    /// A simple list of instance IDs, as used in GameMaker 2024.13 and above.
+    /// </summary>
+    public class InstanceIDList : UndertaleObject
+    {
+        public UndertaleObservableList<int> InstanceIDs { get; set; } = new();
+
+        public void Serialize(UndertaleWriter writer)
+        {
+            writer.Write(InstanceIDs.Count);
+            foreach (int id in InstanceIDs)
+            {
+                writer.Write(id);
+            }
+        }
+
+        public void Unserialize(UndertaleReader reader)
+        {
+            int count = reader.ReadInt32();
+            InstanceIDs.SetCapacity(count);
+            for (int i = 0; i < count; i++)
+            {
+                InstanceIDs.InternalAdd(reader.ReadInt32());
+            }
         }
     }
 }
