@@ -33,7 +33,6 @@ namespace UndertaleModTests
         public string ScriptErrorMessage => throw new NotImplementedException();
         public string ExePath => throw new NotImplementedException();
         public string ScriptErrorType => throw new NotImplementedException();
-        public bool GMLCacheEnabled => throw new NotImplementedException();
 
         public bool IsAppClosed => throw new NotImplementedException();
 
@@ -44,28 +43,9 @@ namespace UndertaleModTests
         public void EnsureDataLoaded()
         {
         }
-        public async Task<bool> MakeNewDataFile()
+        public bool MakeNewDataFile()
         {
-            await Task.Delay(1); //dummy await
             return true;
-        }
-        public void ReplaceTempWithMain(bool imAnExpertBtw = false)
-        {
-        }
-        public void ReplaceMainWithTemp(bool imAnExpertBtw = false)
-        {
-        }
-        public void ReplaceTempWithCorrections(bool imAnExpertBtw = false)
-        {
-        }
-        public void ReplaceCorrectionsWithTemp(bool imAnExpertBtw = false)
-        {
-        }
-        public void UpdateCorrections(bool imAnExpertBtw = false)
-        {
-        }
-        public void ReapplyProfileCode()
-        {
         }
         public void InitializeScriptDialog()
         {
@@ -86,6 +66,11 @@ namespace UndertaleModTests
             Console.WriteLine(message);
         }
 
+        public void ScriptWarning(string message)
+        {
+            Console.WriteLine($"Warning: {message}");
+        }
+
         public bool ScriptQuestion(string message)
         {
             Console.WriteLine(message);
@@ -94,10 +79,6 @@ namespace UndertaleModTests
         public void ScriptOpenURL(string url)
         {
             Console.WriteLine("Open: " + url);
-        }
-        public void NukeProfileGML(string codeName)
-        {
-            Console.WriteLine("NukeProfileGML(): " + codeName);
         }
         public void UpdateProgressBar(string message, string status, double progressValue, double maxValue)
         {
@@ -178,30 +159,6 @@ namespace UndertaleModTests
         {
             Console.Write("SetUMTConsoleText(): " + message);
         }
-        public void ReplaceTextInGML(string codeName, string keyword, string replacement, bool caseSensitive = false, bool isRegex = false, GlobalDecompileContext context = null, IDecompileSettings settings = null)
-        {
-            Console.Write("ReplaceTextInGML(): " + codeName + ", " + keyword + ", " + replacement + ", " + caseSensitive.ToString() + ", " + isRegex.ToString() + ", " + context?.ToString() + ", " + settings?.ToString());
-        }
-        public void ReplaceTextInGML(UndertaleCode code, string keyword, string replacement, bool caseSensitive = false, bool isRegex = false, GlobalDecompileContext context = null, IDecompileSettings settings = null)
-        {
-            Console.Write("ReplaceTextInGML(): " + code.ToString() + ", " + keyword + ", " + replacement + ", " + caseSensitive.ToString() + ", " + isRegex.ToString() + ", " + context?.ToString() + ", " + settings?.ToString());
-        }
-        public void ImportGMLString(string codeName, string gmlCode, bool doParse = true, bool checkDecompiler = false)
-        {
-            Console.Write("ImportGMLString(): " + codeName + ", " + gmlCode + ", " + doParse.ToString());
-        }
-        public void ImportASMString(string codeName, string gmlCode, bool doParse = true, bool nukeProfile = true, bool checkDecompiler = false)
-        {
-            Console.Write("ImportASMString(): " + codeName + ", " + gmlCode + ", " + doParse.ToString());
-        }
-        public void ImportGMLFile(string fileName, bool doParse = true, bool checkDecompiler = false, bool throwOnError = false)
-        {
-            Console.Write($"ImportGMLFile(): \"{fileName}\", {doParse}, {checkDecompiler}, {throwOnError}");
-        }
-        public void ImportASMFile(string fileName, bool doParse = true, bool nukeProfile = true, bool checkDecompiler = false, bool throwOnError = false)
-        {
-            Console.Write($"ImportASMFile(): \"{fileName}\", {doParse}, {nukeProfile}, {checkDecompiler}, {throwOnError}");
-        }
 
         public void SetFinishedMessage(bool isFinishedMessageEnabled)
         {
@@ -235,19 +192,6 @@ namespace UndertaleModTests
             await Task.Delay(1); //dummy await
         }
 
-        public async Task<bool> GenerateGMLCache(GlobalDecompileContext decompileContext = null, object dialog = null, bool clearGMLEditedBefore = false)
-        {
-            Console.WriteLine(String.Format("GenerateGMLCache(): *decompileContext*{0}, *dialog*{1}, {2}",
-                                            decompileContext is null ? " (null)" : "",
-                                            dialog is null ? " (null)" : "",
-                                            clearGMLEditedBefore.ToString().ToLower())
-                              );
-
-            await Task.Delay(1); //dummy await
-
-            return false;
-        }
-
         protected async Task<object> RunScript(string path)
         {
             string scriptpath = Path.Combine("../../../UndertaleModTool/Scripts/Builtin Scripts/", path);
@@ -271,6 +215,11 @@ namespace UndertaleModTests
         }
 
         public string PromptChooseDirectory()
+        {
+            throw new NotImplementedException();
+        }
+
+        public string PromptSaveFile(string defaultExt, string filter)
         {
             throw new NotImplementedException();
         }
@@ -309,17 +258,19 @@ namespace UndertaleModTests
         {
             throw new NotImplementedException();
         }
-        public bool DummyBool()
+
+        public Task ClickableSearchOutput(string title, string query, int resultsCount, IOrderedEnumerable<KeyValuePair<string, List<(int lineNum, string codeLine)>>> resultsDict, bool showInDecompiledView, IOrderedEnumerable<string> failedList = null)
         {
-            return true;
+            throw new NotImplementedException();
         }
 
-        public void DummyVoid()
+        public Task ClickableSearchOutput(string title, string query, int resultsCount, IDictionary<string, List<(int lineNum, string codeLine)>> resultsDict, bool showInDecompiledView, IEnumerable<string> failedList = null)
         {
+            throw new NotImplementedException();
         }
-        public string DummyString()
+
+        public void ChangeSelection(object newSelection, bool inNewTab = false)
         {
-            return "";
         }
 
         public Task ClickableSearchOutput(string title, string query, int resultsCount, IOrderedEnumerable<KeyValuePair<string, List<(int lineNum, string codeLine)>>> resultsDict, bool showInDecompiledView, IOrderedEnumerable<string> failedList = null)
