@@ -18,12 +18,16 @@ namespace UndertaleModTool
             Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "UndertaleModTool");
         public static string ProfilesFolder = Path.Combine(AppDataFolder, "Profiles");
 
+        /// <summary>
+        /// Whether file associations settings should be prompted for on startup.
+        /// </summary>
+        public static bool ShouldPromptForAssociations { get; set; } = false;
+
         public string Version { get; set; } = MainWindow.Version;
         public string GameMakerStudioPath { get; set; } = "%appdata%\\GameMaker-Studio";
         public string GameMakerStudio2RuntimesPath { get; set; } = "%ProgramData%\\GameMakerStudio2\\Cache\\runtimes";
         public bool AssetOrderSwappingEnabled { get; set; } = false;
         public bool ProfileModeEnabled { get; set; } = false;
-        public bool UseGMLCache { get; set; } = false;
         public bool ProfileMessageShown { get; set; } = false;
         public bool AutomaticFileAssociation { get; set; } = true;
         public bool TempRunMessageShow { get; set; } = true;
@@ -56,11 +60,13 @@ namespace UndertaleModTool
         public double GlobalGridThickness { get; set; } = 1;
         public bool GridThicknessEnabled { get; set; } = false;
 
+        public string TransparencyGridColor1 { get; set; } = "#FF666666";
+        public string TransparencyGridColor2 { get; set; } = "#FF999999";
+
         public bool EnableDarkMode { get; set; } = false;
         public bool ShowDebuggerOption { get; set; } = false;
         public DecompilerSettings DecompilerSettings { get; set; }
-        public string InstanceIdPrefix { get; set; } = "inst_";
-
+        public const string DefaultInstanceIdPrefix = "inst_";
         public byte FunctionColor_0 { get; set; } = 255;
         public byte FunctionColor_1 { get; set; } = 184;
         public byte FunctionColor_2 { get; set; } = 113;
@@ -76,8 +82,14 @@ namespace UndertaleModTool
         public byte LocalColor_0 { get; set; } = 255;
         public byte LocalColor_1 { get; set; } = 248;
         public byte LocalColor_2 { get; set; } = 153;
+        public string InstanceIdPrefix { get; set; } = DefaultInstanceIdPrefix;
 
-        public static Settings Instance;
+        public bool ShowNullEntriesInResourceTree { get; set; } = false;
+
+        public WindowPlacementExtensions.WindowPlacement? MainWindowPlacement { get; set; } = null;
+        public bool RememberWindowPlacements { get; set; } = false;
+
+        public static Settings Instance { get; private set; }
 
         public static JsonSerializerOptions JsonOptions = new()
         {
@@ -104,6 +116,9 @@ namespace UndertaleModTool
                     // No settings JSON exists, so make a new one
                     _ = new Settings() { DecompilerSettings = existingDecompilerSettings ?? new() };
                     Save();
+
+                    // This is theoretically a first bootup, so prompt for file associations
+                    ShouldPromptForAssociations = true;
                     return;
                 }
 
@@ -117,6 +132,19 @@ namespace UndertaleModTool
                 {
                     changed = true;
                     // TODO: When necessary, account for any version upgrades
+                }
+
+                // Use existing decompiler settings (from last settings instance)
+                if (existingDecompilerSettings is not null)
+                    Instance.DecompilerSettings = existingDecompilerSettings;
+
+                // If no settings were supplied at all, generate a new one (can be caused from downgrading)
+                Instance.DecompilerSettings ??= new();
+
+                // Auto-remove "argument{0}" syntax (become "arg{0}" by default)
+                if (Instance.DecompilerSettings.UnknownArgumentNamePattern == "argument{0}")
+                {
+                    Instance.DecompilerSettings.UnknownArgumentNamePattern = "arg{0}";
                 }
 
                 // Use existing decompiler settings (from last settings instance)
@@ -175,13 +203,7 @@ namespace UndertaleModTool
 
         // Inner settings used to store values that we don't have any business reimplementing
         [JsonIgnore]
-        private DecompileSettings InnerSettings { get; } = new DecompileSettings()
-        {
-            UnknownArgumentNamePattern = "argument{0}",
-            RemoveSingleLineBlockBraces = true,
-            EmptyLineAroundBranchStatements = true,
-            EmptyLineBeforeSwitchCases = true
-        };
+        private DecompileSettings _innerSettings;
 
         /// <summary>
         /// Indentation style being used for decompilation.
@@ -202,34 +224,55 @@ namespace UndertaleModTool
         }
 
         // Interface implementation (passes through to inner settings instance)
-        public bool UseSemicolon { get => InnerSettings.UseSemicolon; set => InnerSettings.UseSemicolon = value; }
-        public bool UseCSSColors { get => InnerSettings.UseCSSColors; set => InnerSettings.UseCSSColors = value; }
-        public bool PrintWarnings { get => InnerSettings.PrintWarnings; set => InnerSettings.PrintWarnings = value; }
-        public bool MacroDeclarationsAtTop { get => InnerSettings.MacroDeclarationsAtTop; set => InnerSettings.MacroDeclarationsAtTop = value; }
-        public bool EmptyLineAfterBlockLocals { get => InnerSettings.EmptyLineAfterBlockLocals; set => InnerSettings.EmptyLineAfterBlockLocals = value; }
-        public bool EmptyLineAroundEnums { get => InnerSettings.EmptyLineAroundEnums; set => InnerSettings.EmptyLineAroundEnums = value; }
-        public bool EmptyLineAroundBranchStatements { get => InnerSettings.EmptyLineAroundBranchStatements; set => InnerSettings.EmptyLineAroundBranchStatements = value; }
-        public bool EmptyLineBeforeSwitchCases { get => InnerSettings.EmptyLineBeforeSwitchCases; set => InnerSettings.EmptyLineBeforeSwitchCases = value; }
-        public bool EmptyLineAfterSwitchCases { get => InnerSettings.EmptyLineAfterSwitchCases; set => InnerSettings.EmptyLineAfterSwitchCases = value; }
-        public bool EmptyLineAroundFunctionDeclarations { get => InnerSettings.EmptyLineAroundFunctionDeclarations; set => InnerSettings.EmptyLineAroundFunctionDeclarations = value; }
-        public bool EmptyLineAroundStaticInitialization { get => InnerSettings.EmptyLineAroundStaticInitialization; set => InnerSettings.EmptyLineAroundStaticInitialization = value; }
-        public bool OpenBlockBraceOnSameLine { get => InnerSettings.OpenBlockBraceOnSameLine; set => InnerSettings.OpenBlockBraceOnSameLine = value; }
-        public bool RemoveSingleLineBlockBraces { get => InnerSettings.RemoveSingleLineBlockBraces; set => InnerSettings.RemoveSingleLineBlockBraces = value; }
-        public bool CleanupTry { get => InnerSettings.CleanupTry; set => InnerSettings.CleanupTry = value; }
-        public bool CleanupElseToContinue { get => InnerSettings.CleanupElseToContinue; set => InnerSettings.CleanupElseToContinue = value; }
-        public bool CleanupDefaultArgumentValues { get => InnerSettings.CleanupDefaultArgumentValues; set => InnerSettings.CleanupDefaultArgumentValues = value; }
-        public bool CleanupBuiltinArrayVariables { get => InnerSettings.CleanupBuiltinArrayVariables; set => InnerSettings.CleanupBuiltinArrayVariables = value; }
-        public bool CreateEnumDeclarations { get => InnerSettings.CreateEnumDeclarations; set => InnerSettings.CreateEnumDeclarations = value; }
-        public string UnknownEnumName { get => InnerSettings.UnknownEnumName; set => InnerSettings.UnknownEnumName = value; }
-        public string UnknownEnumValuePattern { get => InnerSettings.UnknownEnumValuePattern; set => InnerSettings.UnknownEnumValuePattern = value; }
-        public string UnknownArgumentNamePattern { get => InnerSettings.UnknownArgumentNamePattern; set => InnerSettings.UnknownArgumentNamePattern = value; }
-        public bool AllowLeftoverDataOnStack { get => InnerSettings.AllowLeftoverDataOnStack; set => InnerSettings.AllowLeftoverDataOnStack = value; }
+        public bool UseSemicolon { get => _innerSettings.UseSemicolon; set => _innerSettings.UseSemicolon = value; }
+        public bool UseCSSColors { get => _innerSettings.UseCSSColors; set => _innerSettings.UseCSSColors = value; }
+        public bool PrintWarnings { get => _innerSettings.PrintWarnings; set => _innerSettings.PrintWarnings = value; }
+        public bool MacroDeclarationsAtTop { get => _innerSettings.MacroDeclarationsAtTop; set => _innerSettings.MacroDeclarationsAtTop = value; }
+        public bool EmptyLineAfterBlockLocals { get => _innerSettings.EmptyLineAfterBlockLocals; set => _innerSettings.EmptyLineAfterBlockLocals = value; }
+        public bool EmptyLineAroundEnums { get => _innerSettings.EmptyLineAroundEnums; set => _innerSettings.EmptyLineAroundEnums = value; }
+        public bool EmptyLineAroundBranchStatements { get => _innerSettings.EmptyLineAroundBranchStatements; set => _innerSettings.EmptyLineAroundBranchStatements = value; }
+        public bool EmptyLineBeforeSwitchCases { get => _innerSettings.EmptyLineBeforeSwitchCases; set => _innerSettings.EmptyLineBeforeSwitchCases = value; }
+        public bool EmptyLineAfterSwitchCases { get => _innerSettings.EmptyLineAfterSwitchCases; set => _innerSettings.EmptyLineAfterSwitchCases = value; }
+        public bool EmptyLineAroundFunctionDeclarations { get => _innerSettings.EmptyLineAroundFunctionDeclarations; set => _innerSettings.EmptyLineAroundFunctionDeclarations = value; }
+        public bool EmptyLineAroundStaticInitialization { get => _innerSettings.EmptyLineAroundStaticInitialization; set => _innerSettings.EmptyLineAroundStaticInitialization = value; }
+        public bool OpenBlockBraceOnSameLine { get => _innerSettings.OpenBlockBraceOnSameLine; set => _innerSettings.OpenBlockBraceOnSameLine = value; }
+        public bool RemoveSingleLineBlockBraces { get => _innerSettings.RemoveSingleLineBlockBraces; set => _innerSettings.RemoveSingleLineBlockBraces = value; }
+        public bool CleanupTry { get => _innerSettings.CleanupTry; set => _innerSettings.CleanupTry = value; }
+        public bool CleanupElseToContinue { get => _innerSettings.CleanupElseToContinue; set => _innerSettings.CleanupElseToContinue = value; }
+        public bool CleanupDefaultArgumentValues { get => _innerSettings.CleanupDefaultArgumentValues; set => _innerSettings.CleanupDefaultArgumentValues = value; }
+        public bool CleanupBuiltinArrayVariables { get => _innerSettings.CleanupBuiltinArrayVariables; set => _innerSettings.CleanupBuiltinArrayVariables = value; }
+        public bool CleanupLocalVarDeclarations { get => _innerSettings.CleanupLocalVarDeclarations; set => _innerSettings.CleanupLocalVarDeclarations = value; }
+        public bool CreateEnumDeclarations { get => _innerSettings.CreateEnumDeclarations; set => _innerSettings.CreateEnumDeclarations = value; }
+        public string UnknownEnumName { get => _innerSettings.UnknownEnumName; set => _innerSettings.UnknownEnumName = value; }
+        public string UnknownEnumValuePattern { get => _innerSettings.UnknownEnumValuePattern; set => _innerSettings.UnknownEnumValuePattern = value; }
+        public string UnknownArgumentNamePattern { get => _innerSettings.UnknownArgumentNamePattern; set => _innerSettings.UnknownArgumentNamePattern = value; }
+        public bool AllowLeftoverDataOnStack { get => _innerSettings.AllowLeftoverDataOnStack; set => _innerSettings.AllowLeftoverDataOnStack = value; }
+
+        public DecompilerSettings()
+        {
+            RestoreDefaults();
+        }
+
+        /// <summary>
+        /// Restores default values for all decompiler settings.
+        /// </summary>
+        public void RestoreDefaults()
+        {
+            _innerSettings = new DecompileSettings()
+            {
+                UnknownArgumentNamePattern = "arg{0}",
+                RemoveSingleLineBlockBraces = true,
+                EmptyLineAroundBranchStatements = true,
+                EmptyLineBeforeSwitchCases = true
+            };
+            IndentStyle = IndentStyleKind.FourSpaces;
+        }
 
         /// <inheritdoc/>
         public bool TryGetPredefinedDouble(double value, [MaybeNullWhen(false)] out string result, out bool isResultMultiPart)
         {
             // Pass through to inner settings instance, which has some predefined values already
-            return InnerSettings.TryGetPredefinedDouble(value, out result, out isResultMultiPart);
+            return _innerSettings.TryGetPredefinedDouble(value, out result, out isResultMultiPart);
         }
     }
 }
