@@ -34,6 +34,16 @@ namespace UndertaleModTool
             if (Settings.Instance.EnableDarkMode)
                 MainWindow.SetDarkTitleBarForWindow(this, true, false);
         }
+
+        private void RestoreButton_Click(object sender, RoutedEventArgs e)
+        {
+            Settings.Instance.DecompilerSettings.RestoreDefaults();
+            Settings.Instance.InstanceIdPrefix = Settings.DefaultInstanceIdPrefix;
+
+            // Force all bindings to be updated
+            DataContext = null;
+            DataContext = Settings.Instance;
+        }
     }
 
     [ValueConversion(typeof(DecompilerSettings.IndentStyleKind), typeof(string))]
@@ -41,16 +51,17 @@ namespace UndertaleModTool
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            switch ((DecompilerSettings.IndentStyleKind)value)
+            if (value is not DecompilerSettings.IndentStyleKind kind)
             {
-                case DecompilerSettings.IndentStyleKind.FourSpaces:
-                    return "4 spaces";
-                case DecompilerSettings.IndentStyleKind.TwoSpaces:
-                    return "2 spaces";
-                case DecompilerSettings.IndentStyleKind.Tabs:
-                    return "Tabs";
+                return null;
             }
-            throw new Exception("Unknown indent style kind");
+            return kind switch
+            {
+                DecompilerSettings.IndentStyleKind.FourSpaces => "4 spaces",
+                DecompilerSettings.IndentStyleKind.TwoSpaces => "2 spaces",
+                DecompilerSettings.IndentStyleKind.Tabs => "Tabs",
+                _ => throw new Exception("Unknown indent style kind")
+            };
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
